Deleting Art Fix (#45)
D.C. changed the trigger to delete art when the number of artists is less than or equal to one.

Jacob found a typo in a /DeleteArt fetch request.

DC and Jacob made it so that the delete button calls /deleteArt rather than /deleteContributingArtist when the user is an admin.